Added in the fancybox plugin and added the ability to delete posts

+++ MayaJapan/Controllers/ImageGroupsController.cs
@@ -67,30 +67,28 @@ namespace MayaJapan.Controllers
                 {
                     imageUrlSet = this.stripImageUrl(unformattedUrlSet);
 
+                    imageGroup.ImageValue = new List<Image>();
+
+                    foreach (var image in imageUrlSet)
+                    {
+
+                        Image imageEntity = new Image();
+                        imageEntity.ImageUrl = image.Trim();
+
+                        ImagesController.Create(imageEntity);
+
+                        imageGroup.ImageValue.Add(imageEntity);
+
+                    }
                 }
 
-                imageGroup.ImageValue = new List<Image>();
-                
-                foreach (var image in imageUrlSet)
-                {
 
-                    Image imageEntity = new Image();
-                    imageEntity.ImageUrl = image.Trim();
-
-                    ImagesController.Create(imageEntity);
-
-                    imageGroup.ImageValue.Add(imageEntity);
-
-                }
 
                 imageGroup.Date = DateTime.Now;
-                try {
-                    db.ImageGroups.Add(imageGroup);
-                    db.SaveChanges();
-                } catch(Exception e)
-                {
-                    var apple = e;
-                }
+
+                db.ImageGroups.Add(imageGroup);
+                db.SaveChanges();
+
                 return RedirectToAction("Index");
             }
 
@@ -135,14 +133,31 @@ namespace MayaJapan.Controllers
             return View(imageGroup);
         }
 
+        public ActionResult Delete(int? id)
+        {
+            if(id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+
+            var imageGroup = db.ImageGroups.Find(id.Value);
+
+            if(imageGroup == null)
+            {
+                return HttpNotFound();
+            }
+
+            return View(imageGroup);
+        }
+
         public void Clear()
         {
             var imageGroups = db.ImageGroups.ToArray();
-            foreach(var imageGroup in imageGroups)
+            foreach (var imageGroup in imageGroups)
             {
                 var images = imageGroup.ImageValue.ToArray();
 
-                foreach(var image in images)
+                foreach (var image in images)
                 {
                     db.Images.Remove(image);
                 }
@@ -152,21 +167,34 @@ namespace MayaJapan.Controllers
             db.SaveChanges();
         }
 
+        [HttpPost]
         // POST: ImageGroups/Delete/5
-        public ActionResult Delete(int id)
+        public ActionResult Delete([Bind(Include = "Id")] ImageGroup imageGroup)
         {
-            ImageGroup imageGroup = db.ImageGroups.Find(id);
+            var password = Request["Password"] != null ? Request["Password"] : null;
 
-            var images = imageGroup.ImageValue.ToArray();
+            var realImageGroup = db.ImageGroups.Find(imageGroup.ID);
 
-            foreach(var image in images)
+            if(password == null)
             {
-                db.Images.Remove(image);
-            }
+                return View(imageGroup);
+            } else if(password != Properties.Resources.Password)
+            {
+                return new HttpUnauthorizedResult();
+            } else
+            {
+                var images = realImageGroup.ImageValue.ToArray();
 
-            db.ImageGroups.Remove(imageGroup);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+                foreach (var image in images)
+                {
+                    db.Images.Remove(image);
+                }
+
+                db.ImageGroups.Remove(realImageGroup);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            
         }
 
         protected override void Dispose(bool disposing)
